rename #78: queries 폴더를 api 폴더 내부로 이동
